Fixed issues due to the assumption of api and app domain being the same - Also moved framework/runtime tests to v1 only place
